:recycle:Refactor: traditions 엔드포인트 추가 및 PageResponse 위치 변경
