+++ src/main/java/com/yfive/gbjs/global/config/SecurityConfig.java
@@ -3,9 +3,13 @@
  */
 package com.yfive.gbjs.global.config;
 
+import com.yfive.gbjs.global.config.jwt.JwtFilter;
+import com.yfive.gbjs.global.config.jwt.JwtTokenProvider;
+import com.yfive.gbjs.global.security.CustomOAuth2UserService;
+import com.yfive.gbjs.global.security.OAuth2LoginSuccessHandler;
 import java.util.Arrays;
 import java.util.List;
-
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -19,13 +23,6 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-
-import com.yfive.gbjs.global.config.jwt.JwtFilter;
-import com.yfive.gbjs.global.config.jwt.JwtTokenProvider;
-import com.yfive.gbjs.global.security.CustomOAuth2UserService;
-import com.yfive.gbjs.global.security.OAuth2LoginSuccessHandler;
-
-import lombok.RequiredArgsConstructor;
 
 @Configuration
 @EnableWebSecurity
@@ -81,7 +78,8 @@ public class SecurityConfig {
                         "/api/seals/**",
                         "/api/spots/**",
                         "/api/courses/**",
-                        "/api/users/**")
+                        "/api/users/**",
+                        "/api/traditions/**")
                     .permitAll()
                     // H2 콘솔
                     .requestMatchers("/h2-console/**")
@@ -95,9 +93,9 @@ public class SecurityConfig {
                 oauth2
                     .userInfoEndpoint(
                         userInfo -> userInfo.userService(oauth2UserService) // 사용자 정보 처리
-                        )
+                    )
                     .successHandler(customSuccessHandler) // 로그인 성공 처리
-            );
+        );
 
     return http.build();
   }
